http://www.elvastower.com/forums/index.php?/topic/13199-invalid-operation-exceptions/ Use 16bit/short indexes everywhere.
git-svn-id: http://svn.uktrainsim.com/svn/openrails/trunk@431 a9f3cfff-6c41-7446-b0ca-052bd761e6eb

+++ Source/RunActivity/Popups/Window.cs
@@ -189,12 +189,12 @@ namespace ORTS.Popups
 			}
 			if (WindowIndexBuffer == null)
 			{
-				var indexData = new int[] {
+				var indexData = new short[] {
 					0, 4, 1, 5, 2, 6, 3, 7,
 					11, 6, 10, 5, 9, 4, 8,
 					12, 9, 13, 10, 14, 11, 15,
 				};
-				WindowIndexBuffer = new IndexBuffer(graphicsDevice, typeof(int), indexData.Length, BufferUsage.WriteOnly);
+				WindowIndexBuffer = new IndexBuffer(graphicsDevice, typeof(short), indexData.Length, BufferUsage.WriteOnly);
 				WindowIndexBuffer.SetData(indexData);
 			}
